feat: Config partial access security

+++ backend/src/main/java/com/gabrielrodrigues/encurtador_de_links/controllers/LinkController.java
@@ -15,6 +15,7 @@ import com.gabrielrodrigues.encurtador_de_links.services.UserService;
 import jakarta.annotation.security.PermitAll;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import org.apache.tomcat.util.http.parser.Authorization;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -39,38 +40,15 @@ public class LinkController {
         this.userService = userService;
     }
 
+    @AdminAccess
     @GetMapping("/ping")
     public String pong() {
         return "pong";
     }
 
-    @UserAccess
-    @PostMapping("/shorten")
-    public ResponseEntity<ResponseShortShortenLinkDto> shortenLink(@RequestBody ShortenRequestDto shortenRequest, HttpServletRequest request) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (!(authentication.getPrincipal() instanceof Jwt jwt))
-            throw new TokenNotFoundException("O token jwt não foi encontrado na requisição!");
-
-        Long userId = jwt.getClaim("id");
-
-        return ResponseEntity
-                .status(HttpStatus.CREATED)
-                .body(this.linkService.createShortUrl(shortenRequest.url(), request, userId));
-    }
-
     @AdminAccess
     @GetMapping("/links/{userId}")
     public ResponseEntity<List<ResponseFullShortenLinkDto>> GetLinksByUserId(@PathVariable Long userId, Pageable pageable) {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (!(authentication.getPrincipal() instanceof Jwt jwt))
-            throw new TokenNotFoundException("O token jwt não foi encontrado na requisição!");
-
-        Long userIdJwtToken = jwt.getClaim("id");
-        User user = this.userService.getById(userId);
-
-        if (!Objects.equals(user.getId(), userIdJwtToken))
-            throw new ContentNotAvaliableToTheUserException("Esses dados não pertencem ao usuário logado!");
-
         Page<Link> links = this.linkService.getAllByUserId(userId, pageable);
         List<ResponseFullShortenLinkDto> response = links.stream()
                 .map(link -> new ResponseFullShortenLinkDto(
@@ -108,7 +86,6 @@ public class LinkController {
     @AdminAccess
     @GetMapping("/short/{shortUrl}")
     public ResponseEntity<ResponseFullShortenLinkDto> getByShortUrl(@PathVariable String shortUrl) {
-
         Link link = this.linkService.getByShortUrl(shortUrl);
         ResponseFullShortenLinkDto fullShortenLInkDto = new ResponseFullShortenLinkDto(
                 link.getShortUrl(),
@@ -119,9 +96,43 @@ public class LinkController {
     }
 
     @UserAccess
+    @PostMapping("/shorten")
+    public ResponseEntity<ResponseShortShortenLinkDto> shortenLink(@RequestBody ShortenRequestDto shortenRequest, HttpServletRequest request) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication.getPrincipal() instanceof Jwt jwt))
+            throw new TokenNotFoundException("O token jwt não foi encontrado na requisição!");
+
+        Long userId = jwt.getClaim("id");
+
+        return ResponseEntity
+                .status(HttpStatus.CREATED)
+                .body(this.linkService.createShortUrl(shortenRequest.url(), request, userId));
+    }
+
+    @UserAccess
     @DeleteMapping("/{shortUrl}")
     public ResponseEntity<?> deleteByShortUrl(@PathVariable String shortUrl) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!(authentication.getPrincipal() instanceof Jwt jwt))
+            throw new TokenNotFoundException("O token jwt não foi encontrado na requisição!");
+
+        String role = jwt.getClaim("role");
+        if (!role.equals("ADMIN")) {
+            //Terminar validação
+        }
+
         this.linkService.deleteByShortUrl(shortUrl);
         return ResponseEntity.noContent().build();
     }
+
+//    private void validateUserToken(Authentication authentication) {
+//        if (!(authentication.getPrincipal() instanceof Jwt jwt))
+//            throw new TokenNotFoundException("O token jwt não foi encontrado na requisição!");
+//
+//        Long userIdJwtToken = jwt.getClaim("id");
+//        User user = this.userService.getById(userIdJwtToken);
+//
+//        if (!Objects.equals(authentication.getName(), user.getUsername()))
+//            throw new ContentNotAvaliableToTheUserException("Esses dados não pertencem ao usuário logado!");
+//    }
 }
